Moved asteroid connection logic from View to State
