+++ core/src/com/panic/tdt4240/view/ViewClasses/PlayCardView.java
@@ -45,7 +45,6 @@ public class PlayCardView extends AbstractView{
     private Skin skin;
     public boolean selectTarget = false;
     private Map map;
-    private ArrayList<AsteroidConnection> connections;
     private ShapeRenderer sr;
     private TextButton finishedButton;
 
@@ -141,7 +140,6 @@ public class PlayCardView extends AbstractView{
     private void setUpMap(){
         final ArrayList<Asteroid> asteroids = map.getAsteroids();
         ArrayList<String> vehicles = new ArrayList<>();
-        connections = new ArrayList<>();
 
         for (int i = 0; i < asteroids.size(); i++) {
             vehicles.addAll(asteroids.get(i).getVehicles());
@@ -165,42 +163,8 @@ public class PlayCardView extends AbstractView{
                 };
             });
             for(Asteroid neighbour: asteroids.get(i).getNeighbours()){
-                if(notConnected(asteroids.get(i).getId(), neighbour.getId())){
-                    AsteroidConnection connection = new AsteroidConnection(
-                            new Vector2(asteroids.get(i).getPosition().x *(Gdx.graphics.getWidth() - asteroid.getWidth()) + asteroid.getWidth()/2,
-                                    asteroids.get(i).getPosition().y *(Gdx.graphics.getHeight() - table.getHeight() - asteroid.getHeight()) + table.getHeight()
-                                            + asteroid.getHeight()/2),
-                            new Vector2(neighbour.getPosition().x *(Gdx.graphics.getWidth() - asteroid.getWidth()) + asteroid.getWidth()/2,
-                                    neighbour.getPosition().y *(Gdx.graphics.getHeight() - table.getHeight() - asteroid.getHeight()) + table.getHeight()
-                                            + asteroid.getHeight()/2),
-                            asteroids.get(i).getId(), neighbour.getId());
-                    connections.add(connection);
-                }
+                ((PlayCardState) state).addConnection(asteroids.get(i), neighbour, asteroid.getWidth(), asteroid.getHeight(), table.getHeight());
             }
-        }
-    }
-
-    private boolean notConnected(String startID, String endID){
-        for(AsteroidConnection connection: connections){
-            if(connection.startID.equals(endID) && connection.endID.equals(startID)){
-                return false;
-            }
-            else if(connection.startID.equals(startID) && connection.endID.equals(endID)){
-                return false;
-            }
-        }
-        return true;
-    }
-    private class AsteroidConnection {
-        private Vector2 start;
-        private Vector2 end;
-        private String startID;
-        private String endID;
-        private AsteroidConnection(Vector2 start, Vector2 end, String startID, String endID){
-            this.start = start;
-            this.startID = startID;
-            this.end = end;
-            this.endID = endID;
         }
     }
 
@@ -230,8 +194,9 @@ public class PlayCardView extends AbstractView{
     public void render(){
         renderer.sb.setProjectionMatrix(cam.combined);
         sr.begin(ShapeRenderer.ShapeType.Filled);
-        for(AsteroidConnection connection :connections){
-            sr.rectLine(connection.start, connection.end, 5.0f);
+        ArrayList<Vector2[]> lines = ((PlayCardState) state).getConnections();
+        for(Vector2[] points : lines){
+            sr.rectLine(points[0], points[1], 5.0f);
         }
         sr.end();
         stage.draw();
